+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## WIP
 
 - Move the config version parsing to a VersionLoader implementation
+- Prevent the use of unsafe statements in var queries (e.g. insert/delete)
+- Add a PHP tokenizer abstraction layer
 
 ## [1.0.0-beta13] - 2019-10-31
 [1.0.0-beta13]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta12...1.0.0-beta13